Move exported interface for class to index.ts
