+++ lib/index.ts
@@ -1,12 +1,28 @@
 import "source-map-support/register";
 import * as cdk from "@aws-cdk/core";
 import * as cfn from "@aws-cdk/aws-cloudformation";
+import * as secrets from "@aws-cdk/aws-secretsmanager";
+import * as s3 from "@aws-cdk/aws-s3";
 import {
-  DatadogIntegrationConfig,
   applyDefaultsToConfig,
   DatadogIntegrationConfigWithDefaults,
 } from "./config";
 import { bucketsToString } from "./util";
+
+type DatadogPermissionsLevel = "Full" | "Core";
+export interface DatadogIntegrationConfig {
+  readonly apiKey: secrets.ISecret;
+  readonly externalId: string;
+
+  readonly site?: string;
+  readonly iamRoleName?: string;
+  readonly permissions?: DatadogPermissionsLevel;
+  readonly forwarderName?: string;
+  readonly installDatadogPolicyMacro?: boolean;
+
+  readonly logArchives?: s3.Bucket[] | undefined;
+  readonly cloudTrails?: s3.Bucket[] | undefined;
+}
 
 const DATADOG_AWS_ACCOUNT_ID = "464622532012"; // DO NOT CHANGE!
